Added:Maximum Among three float

+++ src/com/brigelabz/Maximum.java
@@ -11,19 +11,19 @@ public class Maximum {
                       }
 
     // compareTo method check the maximum number among three numbers and return it :-
-    public Integer compareTo() {
+    public float compareTo() {
 
         Scanner scanner = new Scanner(System.in);
 
         System.out.println("Enter first number");
-        Integer number1 = scanner.nextInt();
+        float number1 = scanner.nextFloat();
 
         System.out.println("Enter second number");
-        Integer number2 = scanner.nextInt();
+        float number2 = scanner.nextFloat();
 
         System.out.println("Enter third number");
-        Integer number3 = scanner.nextInt();
-        Integer maximum = number1;
+        float number3 = scanner.nextFloat();
+        float maximum = number1;
 
         if (number1 > number2 && number1 > number3) {
             maximum = number1;
